requirements.txt
streamlit
pandas
duckdb
google-generativeai
pydantic
pytest
sqlglot
